accel-config: Manage idxd_mdev and idxd kernel modules
Two interdependent modules, idxd_mdev and idxd needs to be loaded.
Loading of either loads the other module, however idxd cannot be
unloaded without unloading idxd_mdev. If at start, idxd_mdev is not
loaded, load it and check for idxd. On error, unref modules and context
and exit. At the end, do a reset of driver by unloading idxd_mdev and
reloading it.

Signed-off-by: Ramesh Thomas <ramesh.thomas@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>