accel-config: Remove duplicate movdir64b and enqcmd instructions
There are 2 versions of enqcmd and movdir64b under accel-config test directory.
One in accfg_test.h and one in dsa.h. The version in dsa.h doesn't have the "sfence"
called before the submit instruction and is wrong for DSA. The version in accfg_test.h
is the correct version.

Signed-off-by: Tony Zhu <tony.zhu@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Reviewed-by: Ramesh Thomas <ramesh.thomas@intel.com>