accel-config: Fix idxd modprobe issue with idxd_mdev for user test
User test will clean the idxd module in the memory moving tests. As
idxd_mdev is supported and it depends on idxd, should change the clean
as idxd_mdev. To improve the robust and compatibility to old kernel,
we add the condition to check if the module is loaded and keep idxd
reset.

Signed-off-by: Tony Zhu <tony.zhu@intel.com>
Reviewed-by: Ramesh Thomas <ramesh.thomas@intel.com>