accel-config: Refactor handling of cdev_minor at wq enable/disable
cdev_minor depends on wq being enabled, of type "user" and in non uacce
mode. Update the cached value of cdev_minor when wq state changes by
reading sysfs. If the read fails, a negative value will be stored
indicating it is not valid for the current state of the wq.

Signed-off-by: Ramesh Thomas <ramesh.thomas@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>