add per-wq prs disable tests
when prs is disabled on a WQ, completion record page faults
are recorded in the event log and the driver copies the completion
record to user space, add a flag (0x10) to
generate completion record faults. At present there is no
support (sysfs?) to compare the number of faults processed by
the driver to the number of faults expected to be generated.

add a flag (0x20) for generating page fault on the descriptor list address
within a batch descriptor. When prs is disabled on a WQ, and the device
receives a fault response to the ats request for the descriptor list address,
the device completes the operaton with status code 6 instead of
generating a prs request for the fault address.

as part of the changes above also add a -n flag to generate multiple faults, so
we can test driver event log handling for multiple completion record faults.

add the -e flag that allows for testing fault sequences that can happen
inside of a batch descriptor, these cases are exercised by the test_all
script.

Signed-off-by: Nikhil Rao <nikhil.rao@intel.com>