accel-config/test: add CRC Generation and Copy with CRC Generation support
The CRC Generation operation computes the CRC on memory at the Source
Address.The Copy with CRC Generation operation copies memory from the
Source Address to the Destination Address and computes the CRC on the
data copied. The completion record format for Copy with CRC Generation
is identical to the format for CRC Generation.

Signed-off-by: Tony Zhu <tony.zhu@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>