accel-config/test: Fix descriptor timeout calculation to not overflow
Testing has shown that when using large transfer size, the mwait value is
not working correctly. The dsa_test application will fail due to constant
value converted by compiler to integer and cause the math to overflow on
large values. Explicitly declare constant to UL in order to force type
casting of unsigned long. Also, the calculation should use the passed in
timeout parameter and not the global constant.

Signed-off-by: Ken Buchanan <kenneth.buchanan@intel.com>
Signed-off-by: Tony Zhu <tony.zhu@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>