accel-config/test: Fix dedicated mode wq size 1 failure
Noop operation test of iaa_test fails on dedicated WQ size 1, the root
cause is if the size is 1 and subtract 1, it is 0. So it is looped there.
The correct way is if it is shared mode, the test range is threshold.
If it is dedicated, the test range is wq size, don't need to subtract 1.

Signed-off-by: Li Zhang <li4.zhang@intel.com>
Reviewed-by: Ramesh Thomas <ramesh.thomas@intel.com>