accel-config: Fix a bug in the --help option
CMD --help was not loading the correct man pages because the wrong page
name was being passed to the man command. The short usage help invoked
with "CMD -help" works. However, the "CMD --help" option that displays
the detailed help in the man pages is the one that "accel-config --help"
lists as the way to get help for commands.

Signed-off-by: Ramesh Thomas <ramesh.thomas@intel.com>
Reported-by: Reinette Chatre <reinette.chatre@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>