accel-config/test: Fix bug about finding a mismatched wq
When both DSA wq and IAA wq are enabled, the test code would find
the first available wq for submitting descriptors. If the DSA wq
is prior to IAA wq, the IAA test code would use DSA wq. It's incorrect.
For fixing this bug, jump over the wq with mismatched device type.

Signed-off-by: Rex Zhang <rex.zhang@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>