accel-config/test: Add op mask to check if an op is available
Several operations were removed in IAA2.0 specification. In idxd-config,
keep the related code about removed operations, define op mask for
comparing against the /sys/bus/dsa/devices/iax1/op_cap to check if an
operation is available or not. Whether to do the operation or not
depends on the checking result.

This commit is for defining the operation masks.

Signed-off-by: Rex Zhang <rex.zhang@intel.com>