accel-config/test: Decide whether to do the operation or not
Several operations were removed in IAA2.0 specification. In idxd-config,
keep the related code about removed operations, define op mask for
comparing against the /sys/bus/dsa/devices/iax1/op_cap to check if an
operation is available or not. Whether to do the operation or not
depends on the checking result.

This commit is for deciding whether to do the operation or not depends
on the check result.

Signed-off-by: Rex Zhang <rex.zhang@intel.com>