accel-config: Fixes for pedantic compiler warnings
Debian builds enable the -Wpedantic option that gives warnings for
features that are supported by GCC but prohibited by other compilers.
This project is tied to the GCC compiler and uses several GCC
builti-ins. Provide fixes to warnings that are potential issues.
Several benighn warnings are suppressed using GCC #pragma.

Signed-off-by: Ramesh Thomas <ramesh.thomas@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Reviewed-by: Colin King <colin.king@intel.com>