accel-config/test: Add parameter for specifying dev and wq
About the script iaa_user_test_runner.sh, previously, the only one behavior
is to enable dev and wq firstly than invoke iaa_test, then the iaa_test
will find which dev and wq are enabled automatically.
For enabling the dev and wq outside the script, the script indentify the
enabled dev and wq, then pass the dev and wq into iaa_test as a parameter.

Signed-off-by: Rex Zhang <rex.zhang@intel.com>