accel-config: Remove SET_ERR macro which is error prone
SET_ERR macro was written to help reduce chiecking return value and
exiting in many assignments. However, it is error prone if the order is
not maintained when special handling of the return value is required.
Replace with simple assignment to a variable and checking it for exiting
or special handling.

Signed-off-by: Ramesh Thomas <ramesh.thomas@intel.com>