accel-config/test: disable enabled devices before unloading iaa_crypto
When a idxd device is probed and iaa_crypto will allow the idxd device
to be configured with default values. This is kernel change from patch
"Add support for device/wq defaults". iaa_crypto could not be removed
because this module is used. Before remove it, all enabled dsa/wqs must
be disabled. For the reset of idxd driver, must disable all enabled
devices.

Signed-off-by: Tony Zhu <tony.zhu@intel.com>