test: Use ENQCMD for Betch Testing
The idxd tests are suggested to use syscall write to submit instead of
ENQCMD. But so far the write function cannot handle batch tests.

Use the option '-u' to force the batch tests using ENQCMD.
This is likely a workaround, need to look into the write function of the
cdev supporting the betch function.

Signed-off-by: Yi Sun <yi.sun@intel.com>